Skip build+push based on time instead (#4)
Manifests (for multi-platform images) are not entirely the same as images. As setting annotations on them (instead of labels) to detect necessary (re)builds is complex, this uses the time of the git commit vs. the time of the image build for the decision. Note that the test is not entirely exact, as what is returned is the time for when the image for the current platform was built (not the manifest), but this information is "good enough" as the build is only run from time to time.