Pinpoint version and control Installl (#7)
* Parse a --version and pass further

* Add installation (at specific version) control

This adds the ability to install csharpier locally in a controllable
fashion. When a version is provided, it will be respected when
installing (if relevant) and running. The defaults are still as little
invasive as possible, i.e.:

+ When no version is specified, the installed csharpier or the latest
  Docker image will be used.
+ When a version is specified, it is installed and used.

When to install, and how to run csharpier can be controlled both at the
repository level and at the local level. This is achieved through having
CLI options for repository-wide settings, and recognising environment
variables to overrides those settings, if necessary, at the local
installation.

* Improve log messages