Respect .dockerignore when re-building Docker
This commit adds a .dockerignore file to the repository, and modifies
the logic that is used to determine if a Docker image needs to be
rebuilt. This is to avoid unnecessary rebuilds when only files that are
ignored by the .dockerignore file are changed.

The driving force for this optimisation is that one of the GitHub
actions will, from time to time, generate a useless commit onto the
repository in order to keep it active at GitHub. Without this change,
the generated commit will have the side-effect of rebuilding all Docker
images from time to time, even though the code hasn't changed.