Silence dotnet (#11)
* Add environment vars to silence dotnet

Add environment variables to silence telemetry and the dotnet welcome
message. This makes the output of the image and the pre-commit more
readable.

* Move silencing to a separate function

Since the implementation might use dotnet twice, arrange to move the
silencing mechanism into a separate function that is used from the
different functions that are calling out to dotnet.

* Rewrite ENV setting into a single line

This minimizes the number of layer, but mostly makes the Dockerfile
clearer.